Allow setting/overriding the associated bug id
This adds a {{ id:something }} option that overrides the detected
Bug id. It also allows setting the id for standalone cards so we
can set blocking relationships for those.

This also changes the behaviour of {{ blocks: }} so it does not
require a valid bugzilla id. {{ bug: }} still checks for bug format
validity though.

Couple of small issues were fixed as well

- the Done? flag is not set when the card in question is already
  in the Documentation column.

- cards with custom id are not moved to Done automatically