Improve triage rules
Triage flag is automatically removed from cards in done colums
and from cards for finished bugs (MODIFIED and above).

The same rule applies to adding the triage flag - cards in
done columns should not be getting it.

Bugzilla linked cards follow triaging rules for bugs and watch for:
- status
- targetMilestone
- FutureFeature
- triaged keyword

And standalone cards (with or w/o custom id) follow triaging
rules for cards and watch for:
- status
- target
- score
- dueDate